+++ pygame/controllers-prototype.py
@@ -1,7 +1,8 @@
-import time
 import pygame
 from controllers.GameLoopState import GameLoopState
-from controllers.ControllersManager import ControllersManager
+from controllers.Manager import ControllersManager
+from controllers.Time import *
+from controllers.Input import *
 
 
 def main():
@@ -57,77 +58,6 @@ def main():
 
 		ctrl_mgr.update()
 		game_loop.update()
-
-
-class ControllerBase(object):
-	def __init__(self, input_controllers=tuple()):
-		self._in = input_controllers
-		self._out = None
-
-	def __str__(self):
-		return str(self._out)
-
-	@property
-	def value(self):
-		return self._out
-
-	def update(self):
-		self._out = self._update()
-
-	def _update(self):
-		error_message = f"You have to derive from the {self.__class__.__name__} class and define your own _update() method."
-		raise NotImplementedError(error_message)
-
-
-class ControllerKeysPressed(ControllerBase):
-	def __init__(self):
-		super().__init__()
-		self._out = set()
-
-	def __str__(self):
-		return f"CtrlKeysPressed({','.join(map(str, self._out))})"
-
-	def _update(self):
-		return self._out
-
-	def key_pressed(self, key):
-		self._out.add(key)
-
-	def key_released(self, key):
-		self._out.remove(key)
-
-
-class ControllerKeyValue(ControllerBase):
-	def __init__(self, key):
-		super().__init__()
-		self._key = key
-
-	def _update(self):
-		return self._key
-
-
-class ControllerKeyPressed(ControllerBase):
-	def __init__(self, ctrl_keys_pressed, ctrl_key):
-		super().__init__((ctrl_keys_pressed, ctrl_key))
-
-	def _update(self):
-		return self._in[1].value in self._in[0].value
-
-
-class ControllerTimeTicks(ControllerBase):
-	def _update(self):
-		return time.perf_counter()
-
-
-class ControllerTimeDelta(ControllerBase):
-	def __init__(self, ctrl_time_ticks):
-		super().__init__((ctrl_time_ticks, ))
-		self._last_ticks = 0
-
-	def _update(self):
-		delta = self._in[0].value - self._last_ticks
-		self._last_ticks = self._in[0].value
-		return delta
 
 
 if __name__ == "__main__":
