+++ pygame/controllers/__init__.py
@@ -2,4 +2,7 @@
 """Controllers package"""
 
 from . import GameLoopState
-from . import ControllersManager
+from . import Base
+from . import Manager
+from . import Input
+from . import Time
